Complex fix
-change irq status to 16 bit
-restore LPD CH1 frequency
-set preamble length to 3 bytes
-enable RX timeout again and set it to 50 ms
-process rx_timeout interrupt

+++ CubeIDE/Code/inc/radio.h
@@ -16,7 +16,7 @@ uint8_t rf_tx_packet(void);
 uint8_t rf_start_rx(void);
 void rf_get_rx_packet(void);
 uint8_t rf_get_status(void);
-uint8_t rf_get_irq_status(void);
+uint16_t rf_get_irq_status(void);
 void rf_clear_irq(void);
 void rf_flush_fifo(void);
 uint8_t *get_air_packet_tx(void);
